pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meshtui"
version = "0.1.1"
authors = [
  { name="Emiel Kollof", email="emiel@kollof.nl" },
]
description = "Terminal User Interface for MeshCore companion radios"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Ham Radio",
    "Topic :: Terminals",
]
keywords = ["mesh", "radio", "tui", "terminal", "meshcore", "lora"]
license = "MIT"
license-files = ["LICENSE*"]
dependencies = [
    "meshcore >= 2.1.9",
    "textual >= 0.50.0",
    "requests >= 2.28.0",
    "aiohttp >= 3.9.0",
    "notify2 >= 0.3.1",
    "dbus-python >= 1.2.18",
]

[project.urls]
Homepage = "https://github.com/ekollof/meshtui"
Issues = "https://github.com/ekollof/meshtui/issues"
Repository = "https://github.com/ekollof/meshtui"

[project.scripts]
meshtui = "meshtui.app:main"
meshcore-tcp-proxy = "meshtui.proxy.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "pytest-cov >= 4.0.0",
    "pytest-mock >= 3.10.0",
]
proxy = [
    "pyserial >= 3.5",
    "pyserial-asyncio >= 0.6",
    "pyyaml >= 6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--cov=src/meshtui",
    "--cov-report=term-missing",
    "--cov-report=html",
]
